Fixed TSP model not learning. This will be a checkpoint

+++ src/common/utils.py
@@ -117,7 +117,7 @@ def cal_distance(xy, visiting_seq):
     rolled_seq = np.roll(original_seq, -1, 1)
 
     segments = np.sqrt(((original_seq - rolled_seq) ** 2).sum(-1))
-    distance = segments.sum(2).astype(np.float16)
+    distance = segments.sum(2).astype(np.float32)
     return distance
